MENAMBAHKAN PAGE PIUTANG DAN HUTANG PEMBELIAN
1. Menambahkan page piutang pelanggan
2. Menambahkan page hutang pembelian
3. Menambahkan fitur hutang pembelian pada tabel barang masuk
4. Menambahkan fitur piutang pelanggan pada tabel barang keluar
5. Menambahkan fitur pembayaran untuk piutang pada page piutang pelanggan
6. Menambahkan fitur pembayaran untuk hutang pembelian pada page hutang pembelian

+++ app/Http/Controllers/PemasokController.php
@@ -64,6 +64,12 @@ class PemasokController extends Controller
     public function destroy($id)
     {
         $pemasok = Pemasok::findOrFail($id);
+
+        // CEK APAKAH MASIH DIPAKAI PRODUK
+        if ($pemasok->produk()->count() > 0) {
+            return redirect()->back()->with('error', 'Pemasok tidak bisa dihapus karena masih digunakan oleh produk!');
+        }
+
         $pemasok->delete();
 
         return redirect()->back()->with('success', 'Pemasok berhasil dihapus');
